Added func call statement and fixed some bugs

+++ src/parserScript.ps1
@@ -29,12 +29,25 @@ echo "AST Package name : ".$AST_PACKAGE_NAME."\n"
 
 java -jar $WORKING_DIR/lib/cup_v10k.jar -ast $AST_PACKAGE_NAME -parser $PARSER_NAME -buildtree $WORKING_DIR/spec/mjparser.cup
 
+if (Test-Path $WORKING_DIR\$PACKAGE_REL_PATH\$PARSER_NAME.java){
+    echo "Removing old $PARSER_NAME.java";
+    Remove-Item -Path $WORKING_DIR\$PACKAGE_REL_PATH\$PARSER_NAME.java
+    }
 
-Remove-Item -Path $WORKING_DIR\$PACKAGE_REL_PATH\$PARSER_NAME.java
-Remove-Item -Path $WORKING_DIR\$PACKAGE_REL_PATH\sym.java
+if (Test-Path $WORKING_DIR\$PACKAGE_REL_PATH\sym.java){
+    echo "Removing old sym.java";
+    Remove-Item -Path $WORKING_DIR\$PACKAGE_REL_PATH\sym.java
+    }
 
-Move-Item -Path $pwd\$PARSER_NAME.java -Destination $WORKING_DIR\$PACKAGE_REL_PATH
-Move-Item -Path $pwd\sym.java -Destination $WORKING_DIR\$PACKAGE_REL_PATH
+if (Test-Path $pwd\$PARSER_NAME.java){
+    echo "Moving $PARSER_NAME.java"
+    Move-Item -Path $pwd\$PARSER_NAME.java -Destination $WORKING_DIR\$PACKAGE_REL_PATH
+    }
+
+if (Test-Path $pwd\sym.java){
+    echo "Moving sym.java"
+    Move-Item -Path $pwd\sym.java -Destination $WORKING_DIR\$PACKAGE_REL_PATH
+    }
 
 #vratimo se na prethodni direktorijum
 popd
